safety improvements

+++ src/api/v1/controllers/user.controller.ts
@@ -3,6 +3,8 @@ import { UserService } from "../../../services/user.service.js";
 import { Logger } from "../../../utils/logger.js";
 import { IUser } from "../../../models/user.model.js";
 import { Authentication } from "../../../middleware/auth.middleware.js";
+import { AppError } from "../../../models/error.model.js";
+import { CustomRequest } from "../../../types/common.type.js";
 
 export class UserController {
   constructor(
@@ -18,6 +20,13 @@ export class UserController {
   ): Promise<void> => {
     try {
       const userId = req.params.userId as string;
+      if (userId !== (req as CustomRequest).user.id) {
+        throw new AppError(
+          403,
+          "User",
+          "You are not authorized to update this user"
+        );
+      }
       this.logger.info(`Attempting to get user information`, { userId });
       const result: IUser = await this.userService.getUserById(userId);
       this.logger.info("User data downloaded", { result });
@@ -40,6 +49,13 @@ export class UserController {
     try {
       const code = req.body.code as string;
       const userId = req.params.userId as string;
+      if (userId !== (req as CustomRequest).user.id) {
+        throw new AppError(
+          403,
+          "User",
+          "You are not authorized to update this user"
+        );
+      }
       this.logger.info(`Attempting to join quiz with code ${code}`);
       const result: string = await this.userService.joinQuiz(userId, code);
       this.logger.info(`User joined quiz with code ${code}`);
@@ -63,6 +79,13 @@ export class UserController {
   ): Promise<void> => {
     try {
       const userId: string = req.params.userId;
+      if (userId !== (req as CustomRequest).user.id) {
+        throw new AppError(
+          403,
+          "User",
+          "You are not authorized to update this user"
+        );
+      }
       const sessionId: string = req.params.sessionId;
       this.logger.info(`Attempting get questions for quiz`, { userId });
       const result = await this.userService.getQuestions(sessionId, userId);
@@ -85,6 +108,13 @@ export class UserController {
   ): Promise<void> => {
     try {
       const userId: string = req.params.userId;
+      if (userId !== (req as CustomRequest).user.id) {
+        throw new AppError(
+          403,
+          "User",
+          "You are not authorized to update this user"
+        );
+      }
       const sessionId: string = req.params.sessionId;
       this.logger.info(`Attempting to end test`, { userId, sessionId });
       await this.userService.endQuiz(userId, sessionId);
@@ -100,6 +130,13 @@ export class UserController {
   ): Promise<void> => {
     try {
       const userId: string = req.params.userId;
+      if (userId !== (req as CustomRequest).user.id) {
+        throw new AppError(
+          403,
+          "User",
+          "You are not authorized to update this user"
+        );
+      }
       const oldPassword: string = req.body.oldPassword;
       const newPassword: string = req.body.newPassword;
       this.logger.info("Attempting to change password", { userId });
@@ -117,6 +154,13 @@ export class UserController {
   ): Promise<void> => {
     try {
       const userId: string = req.params.userId;
+      if (userId !== (req as CustomRequest).user.id) {
+        throw new AppError(
+          403,
+          "User",
+          "You are not authorized to update this user"
+        );
+      }
       const password: string = req.body.password;
       this.logger.info("Attempting to delete user", { userId });
       await this.userService.deleteUser(userId, password);
@@ -134,11 +178,29 @@ export class UserController {
   ): Promise<void> => {
     try {
       const userId: string = req.params.userId;
+      if (userId !== (req as CustomRequest).user.id) {
+        throw new AppError(
+          403,
+          "User",
+          "You are not authorized to update this user"
+        );
+      }
       const updatedUser: Omit<IUser, "_id"> = req.body;
       this.logger.info("Attempting to update user", { userId });
-      await this.userService.updateUser(userId, updatedUser);
+      const result = await this.userService.updateUser(userId, updatedUser);
       this.logger.info(`User updated successfully`, { userId });
-      res.status(204).send();
+      if (
+        result ===
+        "Email changed, you have to activate your account again. A new email has been sent."
+      ) {
+        this.authenticator.blacklist(req, res, next);
+      }
+      res.status(200).json({
+        success: true,
+        data: {
+          user: result,
+        },
+      });
     } catch (error) {
       next(error);
     }
@@ -150,6 +212,13 @@ export class UserController {
   ): Promise<void> => {
     try {
       const userId: string = req.params.userId;
+      if (userId !== (req as CustomRequest).user.id) {
+        throw new AppError(
+          403,
+          "User",
+          "You are not authorized to update this user"
+        );
+      }
       const sessionId: string = req.params.sessionId;
       this.logger.info(`Attempting to get result`, { userId, sessionId });
       const result: number = await this.userService.getResult(
